pyproject.toml
[project]
name = "plexsubs"
version = "0.1.0"
description = "Plex webhook for downloading subtitles"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "requests>=2.25.0",
    "urllib3>=1.26.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "guessit>=3.0.0",
    "langdetect>=1.0.0",
    "iso639-lang>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.hatch.build.targets.wheel]
packages = ["plexsubs"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "ruff>=0.15.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]
